config.py
import google.generativeai as genai

# Clé API Google Gemini
GOOGLE_API_KEY = "AIzaSyAaU0MmHgWAb-cueHp4xP7xqbz339Upjvk"

# Configuration de l'API Gemini
genai.configure(api_key=GOOGLE_API_KEY)

# Sélection du modèle Gemini 2.0 Flash
MODEL_NAME = "gemini-2.0-flash"

#  Prompt
SYSTEM_PROMPT = (
    "Tu es un chatbot spécialisé en santé mentale. "
    "Ta mission est d'écouter et de donner des conseils empathiques aux utilisateurs. "
    "Tu n'es pas un professionnel de la santé, mais tu peux recommander des ressources fiables. "
    "Réponds avec douceur et encouragement."
    "Tu dois toujours répondre dans la langue de l'utilisateur : "
    "si la question est en français, réponds en français. "
    "Si la question est en anglais, réponds en anglais. "
    "Si l'utilisateur pose des questions urgentes sur la dépression ou le suicide, "
    "recommande-lui immédiatement un professionnel ou une ligne d'aide."
    "Tu réponds toujour de façon empathique, donne des réponses encourageantes, motivantes, gentilles, douces et personnalisées en fonction de l'utilisateur."
)

# Création du modèle Gemini avec le prompt
CHATBOT_PIPELINE = genai.GenerativeModel(MODEL_NAME, system_instruction=SYSTEM_PROMPT)
