requirements.txt
gradio
google-generativeai
langdetect